refactor: move http-api-adapters to internal lib
