+++ packages/frameworks-express/src/index.ts
@@ -152,7 +152,7 @@
 import { Auth } from "@auth/core"
 import type { AuthConfig, Session } from "@auth/core/types"
 import { Request as ExpressRequest, Response as ExpressResponse } from "express"
-import { toWebRequest, toExpressResponse } from "./http-api-adapters.js"
+import { toWebRequest, toExpressResponse } from "./lib"
 
 function ExpressAuthHandler(authConfig: AuthConfig) {
   return async (req: ExpressRequest, res: ExpressResponse) => {
