+++ packages/frameworks-express/src/lib/index.ts
@@ -1,0 +1,1 @@
+export * from "./http-api-adapters"
