+++ packages/frameworks-express/tests/http-api-adapters/request.test.ts
@@ -1,4 +1,4 @@
-import { encodeUrlEncoded, toWebRequest } from "../../src/http-api-adapters"
+import { encodeUrlEncoded, toWebRequest } from "../../src/lib"
 import { Request as ExpressRequest } from "express"
 import supertest from "supertest"
 import express from "express"
