+++ packages/frameworks-express/tests/setupTests.ts
@@ -1,2 +1,11 @@
-globalThis.fetch ??= require("undici").fetch
-globalThis.crypto ??= require("node:crypto").webcrypto
+import { fetch, Headers } from "undici"
+import { webcrypto } from "crypto"
+
+// @ts-expect-error
+globalThis.fetch ??= fetch
+
+// @ts-expect-error
+globalThis.fetch ??= Headers
+
+// @ts-expect-error
+globalThis.crypto ??= webcrypto
